[KB-1] Remove Unused
